drt: merge DrtLegsAnalyser into DrtAnalysisControlerListener
